chore: partially migrate kythe/go/util to proto v2 API (#4541)

+++ kythe/go/util/riegeli/transpose_decoder.go
@@ -291,7 +291,7 @@ func parseTransposeStateMachine(src io.Reader, hdr byteReader, compressionType c
 
 			// End of submessage is encoded as protoSubmessageType.
 			if protoWireType(tag&7) == protoSubmessageType {
-				tag -= protoSubmessageType - protoBytesType
+				tag -= uint64(protoSubmessageType - protoBytesType)
 				subtype = delimitedEndOfSubmessageSubtype
 			}
